refactor: Move events by date method to `EventController` (#10)
